+++ src/watch-run.ts
@@ -1,3 +1,4 @@
+import * as path from 'path';
 import * as webpack from 'webpack';
 
 import * as constants from './constants';
@@ -8,12 +9,19 @@ import { readFile } from './utils';
 /**
  * Make function which will manually update changed files
  */
-export function makeWatchRun(instance: TSInstance) {
+export function makeWatchRun(
+  instance: TSInstance,
+  loader: webpack.loader.LoaderContext
+) {
   // Called Before starting compilation after watch
   const lastTimes = new Map<string, number>();
   const startTime = 0;
 
-  return (compiler: webpack.Compiler, callback: () => void) => {
+  // Save the loader index.
+  const loaderIndex = loader.loaderIndex;
+
+  return (compiler: webpack.Compiler, callback: (err?: Error) => void) => {
+    const promises = [];
     if (instance.loaderOptions.transpileOnly) {
       instance.reportTranspileErrors = true;
     } else {
@@ -27,7 +35,7 @@ export function makeWatchRun(instance: TSInstance) {
         }
 
         lastTimes.set(filePath, date);
-        updateFile(instance, filePath);
+        promises.push(updateFile(instance, filePath, loader, loaderIndex));
       }
 
       // On watch update add all known dts files expect the ones in node_modules
@@ -37,7 +45,7 @@ export function makeWatchRun(instance: TSInstance) {
           filePath.match(constants.dtsDtsxOrDtsDtsxMapRegex) !== null &&
           filePath.match(constants.nodeModules) === null
         ) {
-          updateFile(instance, filePath);
+          promises.push(updateFile(instance, filePath, loader, loaderIndex));
         }
       }
     }
@@ -45,18 +53,52 @@ export function makeWatchRun(instance: TSInstance) {
     // Update all the watched files from solution builder
     if (instance.solutionBuilderHost) {
       for (const filePath of instance.solutionBuilderHost.watchedFiles.keys()) {
-        updateFile(instance, filePath);
+        promises.push(updateFile(instance, filePath, loader, loaderIndex));
       }
     }
 
-    callback();
+    Promise.all(promises)
+      .then(() => callback())
+      .catch(err => callback(err));
   };
 }
 
-function updateFile(instance: TSInstance, filePath: string) {
-  updateFileWithText(
-    instance,
-    filePath,
-    nFilePath => readFile(nFilePath) || ''
-  );
+function updateFile(
+  instance: TSInstance,
+  filePath: string,
+  loader: webpack.loader.LoaderContext,
+  loaderIndex: number
+) {
+  return new Promise<void>((resolve, reject) => {
+    // When other loaders are specified after ts-loader
+    // (e.g. `{ test: /\.ts$/, use: ['ts-loader', 'other-loader'] }`),
+    // manually apply them to TypeScript files.
+    // Otherwise, files not 'preprocessed' by them may cause complication errors (#1111).
+    if (
+      loaderIndex + 1 < loader.loaders.length &&
+      instance.rootFileNames.has(path.normalize(filePath))
+    ) {
+      let request = `!!${path.resolve(__dirname, 'stringify-loader.js')}!`;
+      for (let i = loaderIndex + 1; i < loader.loaders.length; ++i) {
+        request += loader.loaders[i].request + '!';
+      }
+      request += filePath;
+      loader.loadModule(request, (err, source) => {
+        if (err) {
+          reject(err);
+        } else {
+          const text = JSON.parse(source);
+          updateFileWithText(instance, filePath, () => text);
+          resolve();
+        }
+      });
+    } else {
+      updateFileWithText(
+        instance,
+        filePath,
+        nFilePath => readFile(nFilePath) || ''
+      );
+      resolve();
+    }
+  });
 }