+++ test/comparison-tests/otherLoadersWatch/app.ts
@@ -1,0 +1,2 @@
+import { message } from './message';
+console.log(message);