+++ test/comparison-tests/otherLoadersWatch/message.ts
@@ -1,0 +1,2 @@
+BUG
+export const message = 'Hello, world!';